// ngx-showcase/projects/timelanes/src/lib/model/period.interface.ts
import { Annotation, AnnotationInternal } from './annotation.interface';
import { PeriodStyle } from './period-style.interface';

export interface Period {
    start: number;
    end: number;
    style?: PeriodStyle;
    annotation1: Annotation;
    annotation2: Annotation;
}

export interface PeriodInternal extends Period {
    group: string;
    annotation1: AnnotationInternal;
    annotation2: AnnotationInternal;
}